C program that checks if a input is a valid digit

+++ 0x04-more_functions_nested_loops/1-isdigit.c
@@ -8,5 +8,5 @@
 
 int _isdigit(int c)
 {
-	return (((int) c >= 0 && (int) c <= 9) ? 1 : 0);
+	return ((c >= 0 && c <= 9) ? 1 : 0);
 }
